--- src/pages/SelectPage/navbar/Navbar.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Navbar = () => {
-  return <h1>Selectpage</h1>;
-};
-
-export default Navbar;
