+++ src/pages/mbtiResultPage/index.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export const MbtiResultPage = () => {
+  return <div>결과 페이지</div>;
+};
+
+export default MbtiResultPage;
